Specify uid when creating the user via useradd
On my ubuntu24.04, the user created via useradd always has uid=1000.
However, the uid I observed entering chroot is the same as my own uid,
which is not 1000. Therefore, tools like sudo will fail because it
cannot recognize what user it is.

This commit fixes the issue by forcing the uid to be the same as the
outer one when creating the user via useradd.